move recursive-delete method into SamExtensions
